final seccion 13

+++ src/app/api/todos/route.ts
@@ -1,3 +1,4 @@
+import { getUserSessionServer } from '@/auth/actions/auth-actions';
 import prisma from '@/lib/prisma'
 import { NextResponse } from 'next/server'
 import * as yup from 'yup';
@@ -30,10 +31,16 @@ const postSchema = yup.object({
 
 export async function POST(request: Request) {
 
+    const user = await getUserSessionServer();
+
+    if (!user) {
+        return NextResponse.json('No autorizado',{status: 401})
+    }
+
     try {
-        const body = await postSchema.validate(await request.json()) ;
+        const {complete, description} = await postSchema.validate(await request.json()) ;
         
-        const todo = await prisma.todo.create({ data: body });
+        const todo = await prisma.todo.create({ data: {complete, description, userId: user.id} });
         
         return NextResponse.json( todo );
    
@@ -47,9 +54,15 @@ export async function POST(request: Request) {
 
 export async function DELETE(request: Request) {
 
+    const user = await getUserSessionServer();
+
+    if (!user) {
+        return NextResponse.json('No autorizado',{status: 401})
+    }
+
     try {
         
-        await prisma.todo.deleteMany({ where: { complete: true } });
+        await prisma.todo.deleteMany({ where: { complete: true, userId: user.id  } });
         
         return NextResponse.json( 'borrados' );
